# src/interfaces/restful/openapi.yaml
openapi: 3.0.0
info:
  title: My User API
  version: 1.0.0
  description: API per la gestione e registrazione degli utenti
servers:
  - url: http://localhost:3000
    description: Server di Sviluppo
tags:
  - name: Authentication
    description: Operazioni relative alla registrazione e autenticazione
  - name: Admin
    description: Operazioni amministrative sugli utenti
  - name: Testing
    description: Endpoint di test per verificare lo stato del servizio
  - name: User
    description: Operazioni utente autenticato

paths:
  /user/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account in the system using a username and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mario.rossi
                password:
                  type: string
                  example: password123
      responses:
        '201':
          description: User successfully registered.
        '500':
          description: Internal server error.

  /user/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error

  /user/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '204':
          description: Logged out successfully
        '500':
          description: Internal server error

  /user/change-password:
    post:
      tags:
        - User
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '204':
          description: Password changed
        '401':
          description: Invalid credentials
        '404':
          description: User not found
        '500':
          description: Internal server error

  /user/change-username:
    post:
      tags:
        - User
      summary: Change own username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentUsername:
                  type: string
                newUsername:
                  type: string
      responses:
        '204':
          description: Username changed
        '404':
          description: User not found
        '409':
          description: Username taken
        '500':
          description: Internal server error

  /user/admin/reset-password:
    post:
      tags:
        - Admin
      summary: Reset a user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                newPassword:
                  type: string
      responses:
        '204':
          description: Password reset
        '404':
          description: User not found
        '500':
          description: Internal server error

  /user/test:
    get:
      tags:
        - Testing
      summary: Test endpoint
      description: Returns a success message to verify that the service is up and running.
      responses:
        '201':
          description: Test executed successfully.
          content:
            text/plain:
              schema:
                type: string
                example: Test successfully
        '500':
          description: Internal server error.

  /user/{username}:
    delete:
      tags:
        - Admin
      summary: Delete a household user
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: Username of the household user to delete
      responses:
        '204':
          description: User successfully deleted.
        '404':
          description: User not found.
        '400':
          description: Bad request.
        '500':
          description: Internal server error.